组长最新一次提交

+++ crm3/src/main/java/com/sc/controller/XtRoleController.java
@@ -69,6 +69,7 @@ public class XtRoleController {
 	public Message deleteRole(Long roleId) {
 		System.out.println("进入到了删除账户方法！");
 		xtRoleService.deleteXtRole(roleId);
+		xtUserRoleService.deleteXtUserRoleByRoleId(roleId);
 		return new Message("1", "success", "成功");
 	}
 	
@@ -78,6 +79,7 @@ public class XtRoleController {
 		if (ids!=null&&ids.length>0) {
 			for (Long id : ids) {
 				xtRoleService.deleteXtRole(id);
+				xtUserRoleService.deleteXtUserRoleByRoleId(id);
 			}
 		}
 		mav.setViewName("redirect:selectxtrole.do");
@@ -129,7 +131,7 @@ public class XtRoleController {
 		return mav;
 	}
 	
-	@MyLog("赋予权限")
+	@MyLog("赋予角色")
 	@RequestMapping("/giverole.do")
 	public ModelAndView giveRole(ModelAndView mav,Long[] ids,Long roleId) {
 		Date date = new Date();
@@ -163,8 +165,7 @@ public class XtRoleController {
 		List<XtPermissionInfo> list2 = xtPermissionInfoService.selectInfo();
 		//权限选中集合
 		List<XtPermissionInfo> list3 = xtPermissionInfoService.checkInfos(roleId);
-		//权限一些选中属性
-		
+		//权限中一些选中属性
 		for (XtPermissionInfo info1 : list2) {
 				for (XtPermissionInfo obj : list3) {
 					if(obj != null) {
@@ -186,13 +187,14 @@ public class XtRoleController {
 	@RequestMapping("/rolemanage.do")
 	@ResponseBody
 	public Message roleManage(ModelAndView mav,Long roleId,Long[] permissonIds) {
+		System.out.println("选中的权限ID为:"+Arrays.toString(permissonIds));
 		List<XtPermissionInfo> list = xtPermissionInfoService.checkInfos(roleId); //当前角色所有权限
 		XtUserAccount account = (XtUserAccount) SecurityUtils.getSubject().getPrincipal();	//获取账户信息
 		XtUserAccount xtUserAccount = xtUserAccountService.login(account.getUserName());	//获取用户所有信息
 		Date d = new Date();
 		if (permissonIds!=null&&permissonIds.length>0) { 
 			try {
-				if (list != null) {
+				if (list != null&&list.size()>0) {
 					for (XtPermissionInfo info : list) {//删除之前所有的权限
 						if (info.getPermissonId()!=null&&info!=null) {
 							xtPermissionRoleService.deleteXtPermissionRole(info.getPermissonId(), roleId);
